# api.yaml
openapi: 3.0.0
info:
  description: "An API for H.P. Lovecraft's Works' [Electronic Texts](https://www.hplovecraft.com/writings/texts/)."
  version: 0.1.0
  title: Electronic Texts of H.P. Lovecraft’s Works
  termsOfService: http://swagger.io/terms/
  contact:
    email: giorgi.beriashvili@outlook.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
tags:
  - name: texts
    description: Collection of texts
paths:
  /texts:
    get:
      tags:
        - texts
      summary: Get all of the texts
      description: Returns all of the texts
      operationId: getTexts
      requestBody:
        $ref: "#/components/requestBodies/Texts"
      responses:
        "200":
          description: Texts have been returned successfully
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Texts"
        "400":
          description: Invalid status value
  "/texts/{id}":
    get:
      tags:
        - texts
      summary: Find text by ID
      description: Returns a single text object
      operationId: getTextById
      parameters:
        - name: id
          in: path
          description: ID of text to return
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Text has been returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Text"
        "400":
          description: Invalid ID supplied
        "404":
          description: Text not found
externalDocs:
  description: Find out more about the project
  url: https://github.com/GiorgiBeriashvili/lovecraft-api
servers:
  - url: https://localhost:3000/api/v1
components:
  requestBodies:
    Texts:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Texts"
    Entry:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Entry"
    Text:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Text"
  schemas:
    Texts:
      type: array
      items:
        $ref: "#/components/schemas/Entry"
    Entry:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: enum
          description: The category of the text
          enum:
            - Fiction
            - Poetry
            - Essay
            - Letter
        title:
          type: string
        description:
          type: string
    Text:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: enum
          description: The category of the text
          enum:
            - Fiction
            - Poetry
            - Essay
            - Letter
        title:
          type: string
        content:
          type: string